use libfmt for formattig everywhere
