+++ modules/telemetry/telemetry/src/log_sinks/absl_sink.cpp
@@ -12,7 +12,7 @@
 
 namespace heph::telemetry {
 
-AbslLogSink::AbslLogSink() : formatter_([](const LogEntry& l) { return format(l); }) {
+AbslLogSink::AbslLogSink() : formatter_([](const LogEntry& l) { return fmt::format("{}", l); }) {
 }
 
 AbslLogSink::AbslLogSink(ILogSink::Formatter&& f) : formatter_(std::move(f)) {
